requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pyyaml>=6.0
httpx>=0.27.0
pytz>=2024.1
redis>=5.0.0
tiktoken>=0.7.0
pydantic>=2.0.0
